Extracted storage lock to it's own module. (cherry picked from commit 92d1a3207191fe2a9cf007cca8b30b0154670e4a)
